+++ GCECDev/Views/Profile.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+
+using Xamarin.Forms;
+
+namespace GCECDev.Views
+{
+    public partial class Profile : ContentPage
+    {
+        public Profile()
+        {
+            InitializeComponent();
+        }
+    }
+}